# leverage scheduler cache to estimate replicas