--- a/pkg/estimator/server/nodes/filter.go
+++ b/pkg/estimator/server/nodes/filter.go
@@ -2,80 +2,62 @@ package nodes
 
 import (
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	listv1 "k8s.io/client-go/listers/core/v1"
 	schedcorev1 "k8s.io/component-helpers/scheduling/corev1"
 	"k8s.io/component-helpers/scheduling/corev1/nodeaffinity"
 
 	"github.com/karmada-io/karmada/pkg/estimator/pb"
-	"github.com/karmada-io/karmada/pkg/util/helper"
 )
 
 var (
-	tolerationFilterPredicate = func(t *corev1.Taint) bool {
-		// PodToleratesNodeTaints is only interested in NoSchedule and NoExecute taints.
-		return t.Effect == corev1.TaintEffectNoSchedule || t.Effect == corev1.TaintEffectNoExecute
+	unschedulableTaint = corev1.Taint{
+		Key:    corev1.TaintNodeUnschedulable,
+		Effect: corev1.TaintEffectNoSchedule,
 	}
 )
 
-// NodeClaimWrapper is a wrapper that wraps the node claim.
-type NodeClaimWrapper struct {
-	nodeSelector         labels.Selector
-	tolerations          []corev1.Toleration
-	nodeAffinitySelector *nodeaffinity.NodeSelector
-}
-
-// NewNodeClaimWrapper returns a new NodeClaimWrapper.
-func NewNodeClaimWrapper(claim *pb.NodeClaim) (*NodeClaimWrapper, error) {
-	wrapper := &NodeClaimWrapper{}
-	if claim == nil {
-		wrapper.nodeSelector = labels.Everything()
-		return wrapper, nil
+// GetRequiredNodeAffinity returns the parsing result of pod's nodeSelector and nodeAffinity.
+func GetRequiredNodeAffinity(requirements pb.ReplicaRequirements) nodeaffinity.RequiredNodeAffinity {
+	if requirements.NodeClaim == nil {
+		return nodeaffinity.RequiredNodeAffinity{}
 	}
-	if claim.NodeAffinity != nil {
-		selector, err := nodeaffinity.NewNodeSelector(claim.NodeAffinity)
-		if err != nil {
-			return nil, err
-		}
-		wrapper.nodeAffinitySelector = selector
+	pod := &corev1.Pod{
+		Spec: corev1.PodSpec{
+			NodeSelector: requirements.NodeClaim.NodeSelector,
+			Affinity: &corev1.Affinity{
+				NodeAffinity: &corev1.NodeAffinity{
+					RequiredDuringSchedulingIgnoredDuringExecution: requirements.NodeClaim.NodeAffinity,
+				},
+			},
+		},
 	}
-	wrapper.nodeSelector = labels.SelectorFromSet(claim.NodeSelector)
-	wrapper.tolerations = claim.Tolerations
-	return wrapper, nil
-}
-
-// ListNodesByLabelSelector returns nodes that match the node selector.
-func (w *NodeClaimWrapper) ListNodesByLabelSelector(nodeLister listv1.NodeLister) ([]*corev1.Node, error) {
-	nodes, err := nodeLister.List(w.nodeSelector)
-	if err != nil {
-		return nil, err
-	}
-	return nodes, nil
-}
-
-// IsNodeMatched returns whether the node matches all conditions.
-func (w *NodeClaimWrapper) IsNodeMatched(node *corev1.Node) bool {
-	return w.IsNodeAffinityMatched(node) && w.IsNodeSchedulable(node)
+	return nodeaffinity.GetRequiredNodeAffinity(pod)
 }
 
 // IsNodeAffinityMatched returns whether the node matches the node affinity.
-func (w *NodeClaimWrapper) IsNodeAffinityMatched(node *corev1.Node) bool {
-	if w.nodeAffinitySelector == nil {
-		return true
-	}
-	return w.nodeAffinitySelector.Match(node)
+func IsNodeAffinityMatched(node *corev1.Node, affinity nodeaffinity.RequiredNodeAffinity) bool {
+	// Ignore parsing errors for backwards compatibility.
+	match, _ := affinity.Match(node)
+	return match
 }
 
-// IsNodeSchedulable returns whether the node matches the tolerations.
-func (w *NodeClaimWrapper) IsNodeSchedulable(node *corev1.Node) bool {
-	if node.Spec.Unschedulable {
+// IsTolerationMatched returns whether the node matches the tolerations.
+func IsTolerationMatched(node *corev1.Node, tolerations []corev1.Toleration) bool {
+	// If pod tolerate unschedulable taint, it's also tolerate `node.Spec.Unschedulable`.
+	podToleratesUnschedulable := schedcorev1.TolerationsTolerateTaint(tolerations, &unschedulableTaint)
+	if node.Spec.Unschedulable && !podToleratesUnschedulable {
 		return false
 	}
-	if !helper.NodeReady(node) {
-		return false
-	}
-	if _, isUntolerated := schedcorev1.FindMatchingUntoleratedTaint(node.Spec.Taints, w.tolerations, tolerationFilterPredicate); isUntolerated {
+	if _, isUntolerated := schedcorev1.FindMatchingUntoleratedTaint(node.Spec.Taints, tolerations, DoNotScheduleTaintsFilterFunc()); isUntolerated {
 		return false
 	}
 	return true
+}
+
+// DoNotScheduleTaintsFilterFunc returns the filter function that can
+// filter out the node taints that reject scheduling Pod on a Node.
+func DoNotScheduleTaintsFilterFunc() func(t *corev1.Taint) bool {
+	return func(t *corev1.Taint) bool {
+		// PodToleratesNodeTaints is only interested in NoSchedule and NoExecute taints.
+		return t.Effect == corev1.TaintEffectNoSchedule || t.Effect == corev1.TaintEffectNoExecute
+	}
 }