--- a/pkg/util/lifted/scheduler/framework/parallelize/parallelism_test.go
+++ b/pkg/util/lifted/scheduler/framework/parallelize/parallelism_test.go
@@ -18,7 +18,7 @@ limitations under the License.
 // For reference:
 // https://github.com/kubernetes/kubernetes/blob/release-1.23/pkg/scheduler/framework/parallelize/parallelism_test.go
 
-package lifted
+package parallelize
 
 import (
 	"fmt"