**What type of PR is this?**
/kind feature
<!--
Add one of the following kinds:

/kind api-change
/kind bug
/kind cleanup
/kind deprecation
/kind design
/kind documentation
/kind failing-test
/kind feature
/kind flake

-->

**What this PR does / why we need it**:
Leverage scheduler cache to estimate replicas, we will benefit from:

1. Better performance, about 2-4 times elevating.

Before:

```
goos: linux
goarch: amd64
pkg: github.com/karmada-io/karmada/pkg/estimator/server
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/500_nodes_and_10,000_pods_without_affinity_and_tolerations-10                1152        1146023 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_without_affinity_and_tolerations-10                96       12963309 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_taint_and_tolerations-10                      93       11719549 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_node_affinity_and_tolerations-10              80       14192963 ns/op
PASS
ok      github.com/karmada-io/karmada/pkg/estimator/server      20.397s
```

After:

```
goos: linux
goarch: amd64
pkg: github.com/karmada-io/karmada/pkg/estimator/server
cpu: Intel(R) Xeon(R) Platinum 8255C CPU @ 2.50GHz
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/500_nodes_and_10,000_pods_without_affinity_and_tolerations-10                1420         750605 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_without_affinity_and_tolerations-10               176        6645878 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_taint_and_tolerations-10                     178        6971352 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_node_affinity_and_tolerations-10             180        6725716 ns/op
PASS
ok      github.com/karmada-io/karmada/pkg/estimator/server      27.418s
```

2. More flexible. The cahce prepares a lot of important data structures which tend to be useful in the future, e.g., pod affinity and anti-affinity. 

3. More decent. Now the code is distinct and delicate.


**Which issue(s) this PR fixes**:
Fixes #

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required.
-->
```release-note
leverage scheduler cache to estimate replicas
```